Update Started using namespace Added linters Update lint.yaml Added dev dependencies Code refactor File refactor
