+++ .github/workflows/lint.yaml
@@ -3,33 +3,43 @@ name: Lint
 on: [push, pull_request]
 
 jobs:
-  lint:
-    runs-on: ubuntu-latest
+    lint:
+        runs-on: ubuntu-latest
 
-    steps:
-    - name: Checkout code
-      uses: actions/checkout@v3
+        steps:
+            - name: Checkout code
+              uses: actions/checkout@v4
 
-    - name: Set up PHP
-      uses: shivammathur/setup-php@v2
-      with:
-        php-version: '8.3.8'
+            - name: Set up PHP
+              uses: shivammathur/setup-php@v2
+              with:
+                  php-version: "8.3"
 
-    - name: Install linters
-      run: |
-        composer global require "squizlabs/php_codesniffer=*"
-        composer global require phpmd/phpmd
-        composer global require phpstan/phpstan
-        echo "$HOME/.composer/vendor/bin" >> $GITHUB_PATH
+            - name: Set up Node
+              uses: actions/setup-node@v4
+              with:
+                  node-version: "18"
 
-    - name: Run PHP_CodeSniffer
-      run: |
-        phpcs --standard=PSR12 web
+            - name: Install Composer Dependencies
+              run: |
+                  composer install
 
-    - name: Run PHPMD
-      run: |
-        phpmd web text codesize,unusedcode,naming
+            - name: Install Node Dependencies
+              run: |
+                  npm install
 
-    - name: Run PHPStan
-      run: |
-        phpstan analyse web
+            - name: Run PHP_CodeSniffer
+              run: |
+                  composer lint:cs
+
+            - name: Run PHPMD
+              run: |
+                  composer lint:md
+
+            - name: Run PHPStan
+              run: |
+                  phpstan analyse web
+
+            - name: Run Lint Css
+              run: |
+                  npm run lint:css
