+++ .prettierrc.json
@@ -1,0 +1,13 @@
+{
+	"useTabs": true,
+	"tabWidth": 4,
+	"semi": true,
+	"overrides": [
+		{
+			"files": ["web/**/*.css"],
+			"options": {
+				"useTabs": false
+			}
+		}
+	]
+}
