+++ composer.json
@@ -1,0 +1,29 @@
+{
+	"name": "n2n/n2n-solutions",
+	"type": "project",
+	"autoload": {
+		"psr-4": {
+			"Utilities\\": "web/utilities/",
+			"Services\\": "web/services/"
+		},
+		"files": [
+			"web/utilities/Utilities.php"
+		]
+	},
+	"scripts": {
+		"lint": [
+			"@lint:cs",
+			"@lint:md",
+			"@lint:stan"
+		],
+		"lint:cs": "vendor/bin/phpcs --standard=PSR12 web",
+		"lint:md": "vendor/bin/phpmd web text codesize,unusedcode",
+		"lint:stan": "vendor/bin/phpstan analyse web",
+		"fix:cbf": "vendor/bin/phpcbf --standard=PSR12 web"
+	},
+	"require-dev": {
+		"squizlabs/php_codesniffer": "*",
+		"phpmd/phpmd": "^2.15",
+		"phpstan/phpstan": "^1.11"
+	}
+}
