+++ package.json
@@ -1,5 +1,10 @@
 {
+	"scripts": {
+		"lint:css": "prettier --check 'web/**/*.css'",
+		"fix:css": "prettier --write 'web/**/*.css'"
+	},
 	"devDependencies": {
+		"prettier": "^3.3.2",
 		"tailwindcss": "^3.4.4"
 	}
 }
