+++ web/components/sidebard.php
@@ -1,0 +1,1 @@
+<?php
