+++ web/pages/dashboard/index.php
@@ -1,7 +1,7 @@
 <?php
-require_once WEB_PATH . "/utilities/employee.php";
-require_once WEB_PATH . "/utilities/printArray.php";
 
 session_start();
 
-printArray((array) $_SESSION["employee"]);
+use function Utilities\PrintArray;
+
+PrintArray((array) $_SESSION["employee"]);
