--- web/utilities/redirect.php
@@ -1,8 +1,0 @@
-<?php
-
-/** Redirects the user to the specified path. */
-function redirect(string $path): void
-{
-	header("Location: $path");
-	exit;
-}
